Move OpenAI question router into namespace
Since we'll soon be adding a Claude question router, it makes sense to
move the existing question router into the `OpenAI` namespace.
