+++ spec/lib/answer_composition/pipeline/openai/question_router_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe AnswerComposition::Pipeline::QuestionRouter do
+RSpec.describe AnswerComposition::Pipeline::OpenAI::QuestionRouter do # rubocop:disable RSpec/SpecFilePathFormat
   let(:classification_attributes) do
     {
       name: "greetings",
